==> src/2_restring.R <==
# B cell ----
lookup <- c(
  "decreased numbers of b cells and switched memory b cells" = "decreased",
  "normal" = "normal",
  "normal to low" = "borderline",
  "low, decreasing" = "low",
  "nl to low" = "borderline",
  "normal numbers" = "normal",
  "increased" = "high",
  "impaired memory b cell development" = "defective",
  "nl number, decreased memory and switched b cells" = "borderline",
  "very low" = "low",
  "high" = "high",
  "normal, transitional b cell predominance" = "normal",
  "nl number" = "normal",
  "low/variable" = "low",
  "no memory b cells" = "low",
  "low switched memory b cells" = "low",
  "normal numbers of total b cells and subsets" = "normal",
  "sigm+, igd+ cells present, absent sigg+, iga+ and ige+ cells" =  "variable",
  "nl number, poor antibody and memory responses" = "borderline",
  "variable" = "variable",
  "decreased or normal" = "borderline",
  "b cell lymphopenia" = "low",
  "normal slightly reduced" = "borderline",
  "normal b cell numbers, low switched memory b cells" = "borderline",
  "decreased" = "decreased",
  "normal with low memory cells" = "borderline",
  "nl/low" = "borderline",
  "low" = "low",
  "low to normal" = "borderline",
  "hi naïve/transitional, low memory" =  "variable",
  "low class switched b cells" = "low",
  "low, low cd27+ memory b cells poor peripheral b cell tolerance." = "low",
  "defective" = "defective",
  "low b cells" = "low",
  "slightly decreased" = "borderline",
  "normal or low" = "borderline",
  "declining" = "decreased",
  "low or normal" = "borderline",
  "nl number, poor function" = "borderline",
  "normal count, poor function" = "borderline",
  "normal, low memory and isotype switched b cells" = "borderline",
  "normal b cell numbers, but low switched memory. dysregulated germinal centers" = "borderline",
  "abnormal b cell development\r\n(decreased bm b precursors; \r\nhi cd21lo b; decreased memory b cells). decreased response to cd40" = "defective",
  "increased memory b cells" = "high",
  "normal to high" = "borderline",
  "most patients died early, no data" = NA_character_,
  "normal numbers of b cells, low memory b cells" = "borderline",
  "reduced" = "reduced",
  "not assessed" = NA_character_,
  "variably decreased switched memory b cells" = "decreased",
  "normal numbers in childhood usually" = "normal",
  "low or normal numbers of b cells" = "borderline",
  "low, low switched memory b cells" = "low",
  "b cell deficiency variable" = "variable",
  "low number" = "low",
  "immature b cells" = "immature",
  "variably reduced" = "reduced",
  "low switched memory" = "low",
  "decreased switched memory b cells" = "decreased",
  "normal b cell numbers, impaired bcr activation, low memory and isotype switched b cells" = "variable",
  "low b and memory b cells" = "low",
  "increased immature b cells, reduced memory b cells and increased transitional b cells" =  "variable",
  "normal but diminished memory" = "borderline",
  "low b cells, switched and non-switched" = "low",
  "low memory b cells" = "low",
  "normal b cell numbers, decreased  switched memory b cells, and  altered b cell phenotype" = "borderline",
  "low memory b cells, high cd5 b cells" = "borderline",
  "mild pancytopenia" = NA_character_,
  "poor activation, proliferation, motility" = NA_character_,
  "normal,  decreased memory b cells" = "borderline",
  "not reported" = NA_character_,
  "b cell lymphopenia\r\nhi  bm plasmablasts" = "low",
  "reduced memory b cells" = "reduced",
  "variably low switched memory b cells" = "low",
  "slightly low" = "borderline",
  "normal b cell numbers but low memory" = "borderline",
  "normal (decreased memory b cells)" = "borderline",
  "decreased b. b cell development arrested at pro to pre–b stage" = "decreased",
  "low switched and non-switched b cells" = "low",
  "normal,  reduced switched and non-switched memory b cells, baff expression increased" = "borderline",
  "normal numbers of b cells with low memory" = "borderline",
  "nl number, low memory b cells" = "borderline",
  "altered maturation" = "abnormal",
  "nl numbers, low memory b cells" = "borderline",
  "normal, low memory b cells" = "borderline",
  "normal b cell count with low memory b cells and plasmablasts" = "borderline",
  "high plasmablasts" = "high",
  "normal or reduced memory b cells" = "borderline",
  "normal or decreased" = "borderline",
  "decreased memory b cells" = "decreased"
)

df <- df %>%
  mutate(B_cell_count_clean = `B cell count` %>%
           tolower() %>%
           str_trim() %>%
           recode(!!!lookup))







# t cell ----

tcell_lookup <- c(
  "variably low" = "low",
  "normal" = "normal",
  "normal to low" = "mixed",
  "very low" = "low",
  "low-normal cd4" = "low",
  "progressive decrease, abnormal proliferation to mitogens" = "decreased",
  "low cd8, nl cd4, absent mhc i on lymphocytes" = "low",
  "progressive t cell lymphopenia" = "low",
  "nl number, low memory t and treg cells, poor antigen and anti-cd3 proliferation" = "defective",
  "low" = "low",
  "normal numbers, low mitogens" = "defective",
  "nl number, predominant naïve t-cells, poor proliferation" = "defective",
  "nl number, low treg, high cd4, poor function" = "defective",
  "increased dn t cells" = "high",
  "slightly increased dn t cells" = "high",
  "normal, hyperproliferative" = "high",
  "low/variable" = "mixed",
  "normal with high dr" = "defective",
  "normal numbers of t and myeloid cells with decreased nk cells" = "normal",
  "nl number, but low tcr expression" = "mixed",
  "nl to low" = "mixed",
  "nl number, low treg, poor activation and function" = "mixed",
  "absent cd8, nl cd4" = "low",
  "variable t cell counts" = "variable",
  "responses to pha may be decreased" = "defective",
  "decreased or normal, response to pha may be decreased" = "defective",
  "normal t, nk cell numbers.  accumulation of immature nk cells" = "normal",
  "low cd4 cells\r\nabsent mhc ii expression on lymphocytes\r\n\r\n\r\n" = "low",
  "t cell lymphopenia, mostly cd4" = "low",
  "mild cd4 t cell lymphopenia (low naïve cd4 t). normal nk cell numbers" = "mixed",
  "decreased" = "decreased",
  "low, poor proliferation" = "low",
  "nl to low, poor proliferation to antigen" = "defective",
  "normal numbers of cd4+ t cells, variable cd8+ t cells, nk cells normal" = "variable",
  "variable" = "variable",
  "low naïve cd4" = "low",
  "mild decrease" = "decreased",
  "normal, rarely lymphopenia and decreased lymphoproliferation to mitogens and antigens, hypolastic thymus may be present" = "mixed",
  "decreased naive t cells, rtes" = "decreased",
  "decreased or normal, responses to pha may be decreased" = "mixed",
  "low, poor proliferation, few, poorly functioning treg" = "low",
  "profound depletion of cd4+ cells" = "low",
  "increased treg" = "high",
  "reduced" = "reduced",
  "increased activated t cells" = "high",
  "increased t cells" = "high",
  "cd4 and cd8 t cell lymphopenia" = "low",
  "severe t cell lymphopenia at birth, nomrmalised by adulthood" = "low",
  "low cd4, cd8 t cells, increased th17 and th2, poor t cell proliferation" = "mixed",
  "decreased naïve t cells" = "decreased",
  "low or normal" = "mixed",
  "decreased or normal," = "mixed",
  "nl number, absent treg and g/d t cells, impaired tcr activation" = "defective",
  "normal or decreased, tcr activation impaired" = "defective",
  "no memory t cells" = "low",
  "increased activated/ exhausted t cells. increased  production pro-inflammatory cytokines by t cells.  increased mait, tfh cells, nk cells.\r\nnormal treg" = "variable",
  "cd4 and cd8 t cell lymphopenia with low trecs" = "low",
  "normal numbers t, nk cells. increased naïve cd4+ t cells, rte. decreased memory, tfh; cd40l induction.\r\nincreased %’s cd8 t cells; activated t cells" = "variable",
  "nl number, nl/low function" = "defective",
  "nl number, low cytokine production, poor antigen proliferation" = "defective",
  "normal to decreased" = "mixed",
  "increased  memory cd8 t cells, decreased tregs" = "variable",
  "most patients died early, no data" = NA_character_,
  "normal t cell numbers, increased naïve cd4 and cd8 t cell proportions. decreased tcm cd4+ and cd8+ t cells, tem cd8 t cells; mait, tfh cells.  increased th2. low to normal nk cell counts" = "mixed",
  "normal except low th17 and tfh" = "defective",
  "low cd4 and cd8" = "low",
  "not assessed" = NA_character_,
  "progressive decrease" = "decreased",
  "low t cell numbers, impaired t cell activation and proliferation." = "low",
  "5% have < 1500 cd3t cells/ul in neonatal period" = "low",
  "nl to low number" = "mixed",
  "low cd4+, low treg, restricted t cell repertoire, poor tcr signaling" = "low",
  "low t cells- mainly temra and oligoconal" = "low",
  "lymphopenia, decreased mitogen response" = "low",
  "normal or decreased cd4 numbers, t cell dysregulation" = "mixed",
  "low cd4 low recent thymic emigrant cels, poor proliferation to cd3" = "low",
  "nl number, poor proliferation" = "defective",
  "low t cells and trecs" = "low",
  "nl numbers, low antigen specific memory cd4+" = "defective",
  "nl number, poor proliferation to antigen" = "defective",
  "low th17" = "low",
  "mild lymphopenia with decreased cm, em t cells" = "low",
  "pancytopenia with thrombocytopenia most common" = NA_character_,
  "nl number, defective migration, proliferaton" = "defective",
  "low thymic output, normal in vitro proliferation" = "mixed",
  "t cell lymphopenia, reduced naïve t cells" = "low",
  "low ctla4, normal treg counts" = "mixed",
  "normal t cell counts but high tcm, decreased prolfieration" = "mixed",
  "normal total t cells, tcr activation impaired" = "defective",
  "low cd4 t cells" = "low",
  "number decreased, poor response to mitogens and antigens" = "decreased",
  "normal, defective tcr mediated activation" = "defective",
  "mild lymphopenia but mostly intact. increased cd4-cd8- dn t cells, decreased  vδ2+ γδ t, mait and cd56hi nk cells" = "variable",
  "cd8 and cd4 t cells may be decreased" = "mixed",
  "normal cd4 but low treg, low cd8, poor proliferation" = "mixed",
  "low naïve t cells" = "low",
  "decreased t cell proliferation" = "decreased",
  "lymphopenia, lack of trecs, absent proliferation in response to antigens" = "defective",
  "low cd8" = "low",
  "normal numbers t cells, decreased  ctfh cells" = "mixed",
  "mild pancytopenia" = NA_character_,
  "low t cell in infancy" = "low",
  "poor activation, proliferation, motility" = NA_character_,
  "normal numbers" = "normal",
  "normal or low" = "mixed",
  "nl number, poor diversity, poor function" = "defective",
  "low cd4 cells" = "low",
  "nl number, low naïve t cells, restricted repertoire, poor proliferation to cd3" = "mixed",
  "varies from severely decreased (scid) to normal, impaired lymphocyte proliferation" = "variable",
  "not reported" = NA_character_,
  "nl" = "normal",
  "cd4+ t cell, nk cell lymphopenia\r\nlow naïve cd4 and cd8 t cells" = "low",
  "normal or increased activated t cells" = "variable",
  "impaired function" = "defective",
  "t cell proliferation reduced" = "decreased",
  "variable numbers and activation profile" = "variable",
  "low cd4/cd8" = "low",
  "normal (decreased memory t cells)" = "defective",
  "normal numbers of t and nk cells, increased % cd8s" = "variable",
  "poor th1" = "defective",
  "normal overall, th-17 and t-follicular helper cells decreased" = "defective",
  "modestly decreased" = "decreased",
  "normal with th2 skew" = "variable",
  "low, low terminal differentiated  effector memory (temra) cells, low naïve t cells, poor proliferation" = "low",
  "decreased or normal" = "mixed",
  "low mait, cxcr3 th1 t cells" = "low",
  "high dn t cells, low th1, th17 and tfh" = "high",
  "normal lymphocyte numbers, intact t cell function" = "normal"
)

df <- df %>%
  mutate(T_cell_count_clean = `T cell count` %>%
           tolower() %>%
           str_trim() %>%
           recode(!!!tcell_lookup))
