==> src/5_rename.R <==
colnames(df)[colnames(df) == "Major category"] <- "Major category"
colnames(df)[colnames(df) == "Subcategory"] <- "Subcategory"
colnames(df)[colnames(df) == "Disease"] <- "Disease"
colnames(df)[colnames(df) == "Genetic_defect"] <- "Genetic defect"
colnames(df)[colnames(df) == "Inheritance"] <- "Inheritance"
colnames(df)[colnames(df) == "Other affected cells"] <- "Other affected cells"
colnames(df)[colnames(df) == "Associated features"] <- "Associated features"
colnames(df)[colnames(df) == "ICD9"] <- "ICD9"
colnames(df)[colnames(df) == "ICD10"] <- "ICD10"
colnames(df)[colnames(df) == "HPO combined"] <- "HPO IDs"
# colnames(df)[colnames(df) == "HPO (table)"] <- "HPO (table)"
# colnames(df)[colnames(df) == "HPO (subtable)"] <- "HPO (subtable)"
# colnames(df)[colnames(df) == "HPO...18"] <- "HPO...18"
# colnames(df)[colnames(df) == "HPO...19"] <- "HPO...19"
colnames(df)[colnames(df) == "OMIM_ID"] <- "OMIM ID"
colnames(df)[colnames(df) == "GOF/DN"] <- "GOF/DN details"

colnames(df)[colnames(df) == "T cell count"] <- "T cell details"
colnames(df)[colnames(df) == "B cell count"] <- "B cell details"
colnames(df)[colnames(df) == "Immunoglobulin levels"] <- "Ig details"
colnames(df)[colnames(df) == "Neutrophil count"] <- "Neutrophil details"
colnames(df)[colnames(df) == "Inheritance detail"] <- "Inheritance detail"

colnames(df)[colnames(df) == "B_cell_count_clean"] <- "B cell"
colnames(df)[colnames(df) == "T_cell_count_clean"] <- "T cell"
colnames(df)[colnames(df) == "Immunoglobulin_levels_clean"] <- "Ig"
colnames(df)[colnames(df) == "Neutrophil_count_clean"] <- "Neutrophil"
